// view/view-farma/src/app/interface/dto/entity-pb-dto.ts
export interface EntityPbDTO {
        ecode?: string;
        name?: string;
        address?: string;
        numberAddress?: string;
        estado?: string;
        cidade?: string;
        bairro?: string;
        cep?: string;
        phoneNumber?: string;
        cnpj?: string;
        email?: string;
            
}